backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
paho-mqtt>=1.6.1
pydantic>=2.5.0
python-multipart>=0.0.6
aiofiles>=23.2.1
pyserial>=3.5